requirements.txt
python-dateutil
reportlab
gitPython
pytz